+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Url filter with `null` values
+- Improved performance of `search.pages()` by implementing a cache system.
 
 ## [0.11.0] - 2021-01-12
 ### Added